consolidate utils; various other chat service fixes
