+++ parlai/chat_service/services/messenger/message_sender.py
@@ -7,7 +7,7 @@ import json
 import logging
 import requests
 
-import parlai.mturk.core.shared_utils as shared_utils
+import parlai.chat_service.utils.logging as log_utils
 
 MAX_QUICK_REPLIES = 10
 MAX_TEXT_CHARS = 640
@@ -207,7 +207,7 @@ class MessageSender:
                     api_address, params=self.auth_args, json=message
                 )
                 result = response.json()
-        shared_utils.print_and_log(
+        log_utils.print_and_log(
             logging.INFO, '"Facebook response from message send: {}"'.format(result)
         )
         return result
@@ -242,7 +242,7 @@ class MessageSender:
                         api_address, params=self.auth_args, json=payload
                     )
                     result = response.json()
-            shared_utils.print_and_log(
+            log_utils.print_and_log(
                 logging.INFO, '"Facebook response from message send: {}"'.format(result)
             )
             results.append(result)
@@ -256,7 +256,7 @@ class MessageSender:
         message = {'name': name, "profile_picture_url": image_url}
         response = requests.post(api_address, params=self.auth_args, json=message)
         result = response.json()
-        shared_utils.print_and_log(
+        log_utils.print_and_log(
             logging.INFO, '"Facebook response from create persona: {}"'.format(result)
         )
         return result
@@ -268,7 +268,7 @@ class MessageSender:
         api_address = 'https://graph.facebook.com/' + persona_id
         response = requests.delete(api_address, params=self.auth_args)
         result = response.json()
-        shared_utils.print_and_log(
+        log_utils.print_and_log(
             logging.INFO, '"Facebook response from delete persona: {}"'.format(result)
         )
         return result
@@ -317,7 +317,7 @@ class MessageSender:
                     files=filedata,
                 )
         result = response.json()
-        shared_utils.print_and_log(
+        log_utils.print_and_log(
             logging.INFO,
             '"Facebook response from attachment upload: {}"'.format(result),
         )
