+++ parlai/chat_service/services/websocket/run.py
@@ -8,7 +8,7 @@ Websocket Runner.
 """
 from parlai.core.params import ParlaiParser
 from parlai.chat_service.services.websocket.websocket_manager import WebsocketManager
-from parlai.chat_service.core import shared_utils as utils
+import parlai.chat_service.utils.config as config_utils
 
 
 SERVICE_NAME = 'websocket'
@@ -41,7 +41,7 @@ def run(opt):
 if __name__ == '__main__':
     opt = setup_args()
     config_path = opt.get('config_path')
-    config = utils.parse_configuration_file(config_path)
+    config = config_utils.parse_configuration_file(config_path)
     opt.update(config['world_opt'])
     opt['config'] = config
     run(opt)
