+++ parlai/chat_service/tasks/chatbot/config.yml
@@ -10,11 +10,13 @@ overworld: MessengerOverworld
 max_workers: 30
 opt:
   debug: True
-  model: legacy:seq2seq:0
-  model_file: models:convai2/seq2seq/convai2_self_seq2seq_model
-  override:
-    model: legacy:seq2seq:0
-    no_cuda: True
-  no_cuda: True
+  models:
+    legacy_seq2seq:
+      model: legacy:seq2seq:0
+      model_file: models:convai2/seq2seq/convai2_self_seq2seq_model
+      override:
+        model: legacy:seq2seq:0
+        no_cuda: True
+      no_cuda: True
 additional_args:
   page_id: 1 # Configure Your Own Page
